renaming sor folder to lib
