+++ modules/sor/sorV2/beetsHelpers.ts
@@ -1,7 +1,7 @@
 import { GqlPoolMinimal, GqlSorSwapRoute, GqlSorSwapRouteHop } from '../../../schema';
 import { formatFixed } from '@ethersproject/bignumber';
-import { BatchSwapStep, SingleSwap, SwapKind } from './sor-port/types';
-import { NATIVE_ADDRESS, ZERO_ADDRESS } from './sor-port/constants';
+import { BatchSwapStep, SingleSwap, SwapKind } from './lib/types';
+import { NATIVE_ADDRESS, ZERO_ADDRESS } from './lib/constants';
 
 export function mapRoutes(
     swaps: BatchSwapStep[] | SingleSwap,
