+++ modules/sor/sorV2/lib/pools/stable/stableMath.ts
