+++ modules/sor/sorV2/sorV2.service.ts
@@ -9,12 +9,12 @@ import { poolsToIgnore } from '../constants';
 import { AllNetworkConfigsKeyedOnChain } from '../../network/network-config';
 import * as Sentry from '@sentry/node';
 import { getToken } from '../utils';
-import { Swap } from './sor-port/entities/swap';
+import { Swap } from './lib/entities/swap';
 import { Address, parseUnits } from 'viem';
-import { BatchSwapStep, SingleSwap, SwapKind } from './sor-port/types';
-import { sorGetSwapsWithPools } from './sor-port/static';
+import { BatchSwapStep, SingleSwap, SwapKind } from './lib/types';
+import { sorGetSwapsWithPools } from './lib/static';
 import { SwapResultV2 } from './swapResultV2';
-import { TokenAmount } from './sor-port/entities/tokenAmount';
+import { TokenAmount } from './lib/entities/tokenAmount';
 import { poolService } from '../../pool/pool.service';
 import { mapRoutes } from './beetsHelpers';
 import { replaceZeroAddressWithEth } from '../../web3/addresses';
